+++ .github/CODEOWNERS
@@ -6,7 +6,6 @@
 /cmd/                @FerretDB/leads @FerretDB/specops
 /internal/           @FerretDB/leads @FerretDB/specops
 
-/tests/common/       @FerretDB/leads
 /tests/diff/         @FerretDB/leads
 /tests/mongo-tools/  @FerretDB/leads
