+++ tests/diff/nested_arrays_test.go
@@ -20,14 +20,12 @@ import (
 	"github.com/stretchr/testify/require"
 	"go.mongodb.org/mongo-driver/bson"
 	"go.mongodb.org/mongo-driver/mongo"
-
-	"github.com/FerretDB/dance/tests/common"
 )
 
 func TestNestedArrays(t *testing.T) {
 	t.Parallel()
 
-	ctx, db := common.Setup(t)
+	ctx, db := setup(t)
 
 	t.Run("Insert", func(t *testing.T) {
 		t.Parallel()
